App.jsx: reogranize imports, router data, consistent naming
rewrite App.jsx: ordered import,

- reorder imports to have external, components, internal as an order
- use react-router-dom to handle all data, introduce a small data loader
  to do the heavy lifting
- use consistent naming:
  - data -> sheetsData
  - fecth (sic) -> fetchGoogleSheets

App.jsx: revert to loading google sheets from app.

using react loader to load google sheets data is a bad idea as it takes
a fair amount of time that is better used rendering the rest of the App.

we could probably split out the map from App.jsx for this to make sense,
but it looks like a lot more work.

App.jsx: useEffect was missing a dependency on months

detected by linter