bundle size: lazy-load App.jsx
maplibre-gl is a 700kb library, as we use it on the main page we get a
huge bundle size hit.

we use react lazy to split it out to it's own chunk, hopeing that this
can improve performance if not loading the main-page (as we're behind a
route), provide better caching and help with modular updates (i.e. not
updating 1.5mb bundle if not changing anything in the App.jsx file)